% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdrm.r, R/bdrmfit.r
\name{bdrm}
\alias{bdrm}
\alias{bdrm.formula}
\alias{bdrmfit}
\title{Fitting Bayesian Dose-Response Models}
\usage{
bdrm(
  formula,
  data,
  model,
  response = "gaussian",
  binomsize = NULL,
  linfct = NULL,
  fixed = NULL,
  lwr = NULL,
  upr = NULL,
  prior.mu,
  prior.sd,
  atau = 0.001,
  btau = 0.001,
  chains = 4,
  iter = 10000,
  burnin = 9000,
  adapt = 10000,
  startval = NULL
)

\method{bdrm}{formula}(
  formula,
  data,
  model,
  response = "gaussian",
  binomsize = NULL,
  linfct = NULL,
  fixed = NULL,
  lwr = NULL,
  upr = NULL,
  prior.mu,
  prior.sd,
  atau = 0.001,
  btau = 0.001,
  chains = 4,
  iter = 10000,
  burnin = 9000,
  adapt = 10000,
  startval = NULL
)

bdrmfit(
  x,
  y,
  model,
  response,
  binomsize,
  linfct,
  fixed,
  lwr,
  upr,
  prior.mu,
  prior.sd,
  atau,
  btau,
  chains,
  iter,
  burnin,
  adapt,
  startval
)
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} specifying the name of dose and response.}

\item{data}{an optional data frame, list or environment (or object coercible by \code{\link{as.data.frame}} to a data frame) containing the variables in the model. If not found in \code{data}, the variables are taken from \code{environment(formula)}, typically the environment from which \code{bdrm} is called.}

\item{model}{a bdrm model function, specifying the functional dose-response relationship.}

\item{response}{the distribution of the response, either "gaussian", "binomial" or "poisson"}

\item{binomsize}{total number of events if response="binomial"}

\item{linfct}{either a list with a rhs \code{"\link{formula}"} for each of the parameters of the nonlinear dose-response function, conditioning on additional predictor variables, or a list of model matrices, see \code{"\link{model.matrix}"}. The default NULL assumes only an intercept for each of the parameters.}

\item{fixed}{a vector with fixed values for each of the model parameters. When NA a prior distribution is assumed. The default NULL specifies NA for each of the parameters.}

\item{lwr}{a vector with lower bounds for each parameter; NULL assumes -Inf for all parameters.}

\item{upr}{a vector with upper bounds for each parameter; NULL assumes Inf for all parameters.}

\item{prior.mu}{a vector with prior means for each parameter. Also needs to include a value for each fixed parameter as a placeholder.}

\item{prior.sd}{a vector with prior standard deviations for each parameter. Also needs to include a value for each fixed parameter as a placeholder.}

\item{atau}{first parameter of the inverse gamma prior for the residual variance.}

\item{btau}{second parameter of the inverse gamma prior for the residual variance.}

\item{chains}{number of chains for the MCMC algorithm.}

\item{iter}{number of iterations for each chain.}

\item{burnin}{number of warm-up samples.}

\item{adapt}{number of samples in an adaptation phase.}

\item{startval}{a vector with starting values for each parameter. If NULL, the starting values are sampled from the prior distributions.}

\item{x}{a vector with dose levels.}

\item{y}{a vector with response values.}
}
\value{
An object of class \code{"bdrm"} is a list containing at least the following components:
\describe{
 \item{model}{a bdrm model object.}
 \item{response}{distributional assumption about the response.}
 \item{linfct}{a list with model matrices.}
 \item{fixed}{a vector with fixed values for each of the model parameters.}
 \item{lwr}{a vector with lower bounds for each parameter.}
 \item{upr}{a vector with upper bounds for each parameter.}
 \item{prior.mu}{a vector with prior means for each parameter.}
 \item{prior.sd}{a vector with prior standard deviations for each parameter.}
 \item{atau}{first parameter of the inverse gamma prior for the residual variance.}
 \item{btau}{second parameter of the inverse gamma prior for the residual variance.}
 \item{chains}{number of chains.}
 \item{iter}{number of samples from the posterior distribution.}
 \item{burnin}{number of burnin samples.}
 \item{adapt}{number of iterations in an adaptation phase.}
 \item{startval}{a vector with starting values}
 \item{psamples}{an array with samples from the posterior distribution of the model parameters.}
 \item{vsamples}{a matrix with posterior samples of the residual variance.}
 \item{names}{a vector with names for each parameter.}
}
}
\description{
Sampling from the posterior distribution of parameters of a sigmoidal dose-response curve. \code{bdrm} is comparable to function \code{drm} in the drc package, but there are slight differences between the user-interface and of course prior distributions need to be provided.
}
\seealso{
\code{\link[drc]{drm}}
}
\keyword{models}
