% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfunctions.r
\name{logistic}
\alias{logistic}
\title{Logistic Dose-Response Model}
\usage{
logistic(names = c("b1", "b2", "b3", "b4", "b5"))
}
\arguments{
\item{names}{a vector with names for each model parameter}
}
\value{
a list containing at least the following components:
\describe{
  \item{fct}{the logistic function with a dose x and a vector of parameters beta}
  \item{loglik}{the log-likelihood function}
  \item{ss}{sum-of-squares for observed and fitted values}
  \item{p}{number of model parameters}
  \item{names}{a vector with parameter names}
}
}
\description{
Assuming a logistic dose-response relationship in a bdrm model.
}
\details{
The logistic function is defined as 
\deqn{f(x, \beta) = \beta_2 + \frac{\beta_3}{(1 + \exp(-\beta_1 (x-\beta_4))^{\beta_5})}}
}
\keyword{models}
