% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.r
\name{traceplot}
\alias{traceplot}
\title{traceplot for bdrm objects}
\usage{
traceplot(x, which = 1:length(x$names))
}
\arguments{
\item{x}{an object of class bdrm.}

\item{which}{index of the parameter of interest; default are all parameters.}
}
\value{
a ggplot2 plot.
}
\description{
traceplot for bdrm objects
}
