% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfunctions.r
\name{asyreg}
\alias{asyreg}
\title{Asymptotic Regression Model}
\usage{
asyreg(names = c("b1", "b2", "b3"))
}
\arguments{
\item{names}{a vector with names for each model parameter}
}
\value{
a list containing at least the following components:
\describe{
  \item{fct}{the logistic function with a dose x and a vector of parameters beta}
  \item{loglik}{the log-likelihood function}
  \item{ss}{sum-of-squares for observed and fitted values}
  \item{p}{number of model parameters}
  \item{names}{a vector with parameter names}
}
}
\description{
Assuming an asymptotic regression model in bdrm
}
\details{
The asymptotic regression model is defined as 
\deqn{f(x, \beta) = \beta_1 + \beta_2 (1 - \exp(\frac{-x}{\beta_3})) }
}
\keyword{models}
